chatkit-server/requirements.txt
openai-chatkit>=0.1.0
openai-agents>=0.1.0
openai>=1.60.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
